requirements.txt
tensorflow==2.16.1
numpy==1.24.2
Pillow==9.5.0
streamlit==1.26.0
streamlit-drawable-canvas==0.9.2
